+++ exporter/opentelemetry-exporter-otlp-proto-http/src/opentelemetry/exporter/otlp/proto/http/metric_exporter/__init__.py
@@ -262,22 +262,40 @@ class OTLPMetricExporter(MetricExporter, OTLPMetricExporterMixin):
 
             # multiplying by a random number between .8 and 1.2 introduces a +/20% jitter to each backoff.
             backoff_seconds = 2**retry_num * random.uniform(0.8, 1.2)
+            try:
+                resp = self._export(serialized_data, deadline_sec - time())
+                if resp.ok:
+                    return MetricExportResult.SUCCESS
+            except requests.exceptions.RequestException as error:
+                reason = error
+                retryable = isinstance(error, ConnectionError)
+                status_code = None
+            else:
+                reason = resp.reason
+                retryable = _is_retryable(resp)
+                status_code = resp.status_code
+
+            if not retryable:
+                _logger.error(
+                    "Failed to export metrics batch code: %s, reason: %s",
+                    status_code,
+                    reason,
+                )
+                return MetricExportResult.FAILURE
             if (
-                not _is_retryable(resp)
-                or retry_num + 1 == _MAX_RETRYS
+                retry_num + 1 == _MAX_RETRYS
                 or backoff_seconds > (deadline_sec - time())
                 or self._shutdown
             ):
                 _logger.error(
-                    "Failed to export metrics batch code: %s, reason: %s",
-                    resp.status_code,
-                    resp.text,
+                    "Failed to export metrics batch due to timeout, "
+                    "max retries or shutdown."
                 )
                 return MetricExportResult.FAILURE
 
             _logger.warning(
                 "Transient error %s encountered while exporting metrics batch, retrying in %.2fs.",
-                resp.reason,
+                reason,
                 backoff_seconds,
             )
             shutdown = self._shutdown_in_progress.wait(backoff_seconds)