+++ opentelemetry-sdk/src/opentelemetry/sdk/_configuration/__init__.py
@@ -25,7 +25,7 @@ import os
 import warnings
 from abc import ABC, abstractmethod
 from os import environ
-from typing import Any, Callable, Mapping, Sequence, Type, Union
+from typing import Any, Callable, Mapping, Protocol, Sequence, Type, Union
 
 from typing_extensions import Literal
 
@@ -37,7 +37,11 @@ from opentelemetry.environment_variables import (
     OTEL_TRACES_EXPORTER,
 )
 from opentelemetry.metrics import set_meter_provider
-from opentelemetry.sdk._logs import LoggerProvider, LoggingHandler
+from opentelemetry.sdk._logs import (
+    LoggerProvider,
+    LoggingHandler,
+    LogRecordProcessor,
+)
 from opentelemetry.sdk._logs.export import (
     BatchLogRecordProcessor,
     LogRecordExporter,
@@ -48,6 +52,7 @@ from opentelemetry.sdk.environment_variables import (
     OTEL_EXPORTER_OTLP_METRICS_PROTOCOL,
     OTEL_EXPORTER_OTLP_PROTOCOL,
     OTEL_EXPORTER_OTLP_TRACES_PROTOCOL,
+    OTEL_PYTHON_TRACER_CONFIGURATOR,
     OTEL_TRACES_SAMPLER,
     OTEL_TRACES_SAMPLER_ARG,
 )
@@ -58,7 +63,11 @@ from opentelemetry.sdk.metrics.export import (
     PeriodicExportingMetricReader,
 )
 from opentelemetry.sdk.resources import Attributes, Resource
-from opentelemetry.sdk.trace import TracerProvider
+from opentelemetry.sdk.trace import (
+    SpanProcessor,
+    TracerProvider,
+    _TracerConfiguratorT,
+)
 from opentelemetry.sdk.trace.export import BatchSpanProcessor, SpanExporter
 from opentelemetry.sdk.trace.id_generator import IdGenerator
 from opentelemetry.sdk.trace.sampling import Sampler
@@ -105,6 +114,18 @@ ExporterArgsMap = Mapping[
 ]
 
 
+class _ConfigurationExporterSpanProcessorT(Protocol):
+    def __call__(
+        self, span_exporter: SpanExporter, *args, **kwargs
+    ) -> SpanProcessor: ...
+
+
+class _ConfigurationExporterLogRecordProcessorT(Protocol):
+    def __call__(
+        self, exporter: LogRecordExporter, *args, **kwargs
+    ) -> LogRecordProcessor: ...
+
+
 def _import_config_components(
     selected_components: Sequence[str], entry_point_name: str
 ) -> list[tuple[str, Type]]:
@@ -144,6 +165,10 @@ def _get_sampler() -> str | None:
 
 def _get_id_generator() -> str:
     return environ.get(OTEL_PYTHON_ID_GENERATOR, _DEFAULT_ID_GENERATOR)
+
+
+def _get_tracer_configurator() -> str | None:
+    return environ.get(OTEL_PYTHON_TRACER_CONFIGURATOR, None)
 
 
 def _get_exporter_entry_point(
@@ -210,19 +235,29 @@ def _init_tracing(
     sampler: Sampler | None = None,
     resource: Resource | None = None,
     exporter_args_map: ExporterArgsMap | None = None,
+    span_processors: Sequence[SpanProcessor] | None = None,
+    export_span_processor: _ConfigurationExporterSpanProcessorT | None = None,
+    tracer_configurator: _TracerConfiguratorT | None = None,
 ):
     provider = TracerProvider(
         id_generator=id_generator,
         sampler=sampler,
         resource=resource,
+        _tracer_configurator=tracer_configurator,
     )
     set_tracer_provider(provider)
 
     exporter_args_map = exporter_args_map or {}
+    export_processor = export_span_processor or BatchSpanProcessor
+
+    span_processors = span_processors or []
+    for span_processor in span_processors:
+        provider.add_span_processor(span_processor)
+
     for _, exporter_class in exporters.items():
         exporter_args = exporter_args_map.get(exporter_class, {})
         provider.add_span_processor(
-            BatchSpanProcessor(exporter_class(**exporter_args))
+            export_processor(exporter_class(**exporter_args))
         )
 
 
@@ -256,15 +291,24 @@ def _init_logging(
     resource: Resource | None = None,
     setup_logging_handler: bool = True,
     exporter_args_map: ExporterArgsMap | None = None,
+    log_record_processors: Sequence[LogRecordProcessor] | None = None,
+    export_log_record_processor: _ConfigurationExporterLogRecordProcessorT
+    | None = None,
 ):
     provider = LoggerProvider(resource=resource)
     set_logger_provider(provider)
 
     exporter_args_map = exporter_args_map or {}
+    export_processor = export_log_record_processor or BatchLogRecordProcessor
+
+    log_record_processors = log_record_processors or []
+    for log_record_processor in log_record_processors:
+        provider.add_log_record_processor(log_record_processor)
+
     for _, exporter_class in exporters.items():
         exporter_args = exporter_args_map.get(exporter_class, {})
         provider.add_log_record_processor(
-            BatchLogRecordProcessor(exporter_class(**exporter_args))
+            export_processor(exporter_class(**exporter_args))
         )
 
     # silence warnings from internal users until we drop the deprecated Events API
@@ -313,6 +357,27 @@ def _overwrite_logging_config_fns(handler: LoggingHandler) -> None:
     logging.config.fileConfig = wrapper(logging.config.fileConfig)
     logging.config.dictConfig = wrapper(logging.config.dictConfig)
     logging.basicConfig = wrapper(logging.basicConfig)
+
+
+def _import_tracer_configurator(
+    tracer_configurator_name: str | None,
+) -> _TracerConfiguratorT | None:
+    if not tracer_configurator_name:
+        return None
+
+    try:
+        _, tracer_configurator_impl = _import_config_components(
+            [tracer_configurator_name.strip()],
+            "_opentelemetry_tracer_configurator",
+        )[0]
+    except Exception as exc:  # pylint: disable=broad-exception-caught
+        _logger.warning(
+            "Using default tracer configurator. Failed to load tracer configurator, %s: %s",
+            tracer_configurator_name,
+            exc,
+        )
+        return None
+    return tracer_configurator_impl
 
 
 def _import_exporters(
@@ -429,7 +494,14 @@ def _initialize_components(
     id_generator: IdGenerator | None = None,
     setup_logging_handler: bool | None = None,
     exporter_args_map: ExporterArgsMap | None = None,
+    span_processors: Sequence[SpanProcessor] | None = None,
+    export_span_processor: _ConfigurationExporterSpanProcessorT | None = None,
+    log_record_processors: Sequence[LogRecordProcessor] | None = None,
+    export_log_record_processor: _ConfigurationExporterLogRecordProcessorT
+    | None = None,
+    tracer_configurator: _TracerConfiguratorT | None = None,
 ):
+    # pylint: disable=too-many-locals
     if trace_exporter_names is None:
         trace_exporter_names = []
     if metric_exporter_names is None:
@@ -454,6 +526,12 @@ def _initialize_components(
         resource_attributes[ResourceAttributes.TELEMETRY_AUTO_VERSION] = (  # type: ignore[reportIndexIssue]
             auto_instrumentation_version
         )
+    if tracer_configurator is None:
+        tracer_configurator_name = _get_tracer_configurator()
+        tracer_configurator = _import_tracer_configurator(
+            tracer_configurator_name
+        )
+
     # if env var OTEL_RESOURCE_ATTRIBUTES is given, it will read the service_name
     # from the env variable else defaults to "unknown_service"
     resource = Resource.create(resource_attributes)
@@ -464,6 +542,9 @@ def _initialize_components(
         sampler=sampler,
         resource=resource,
         exporter_args_map=exporter_args_map,
+        span_processors=span_processors,
+        export_span_processor=export_span_processor,
+        tracer_configurator=tracer_configurator,
     )
     _init_metrics(
         metric_exporters, resource, exporter_args_map=exporter_args_map
@@ -482,6 +563,8 @@ def _initialize_components(
         resource,
         setup_logging_handler,
         exporter_args_map=exporter_args_map,
+        log_record_processors=log_record_processors,
+        export_log_record_processor=export_log_record_processor,
     )