+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/_incubating/attributes/linux_attributes.py
@@ -15,12 +15,17 @@
 from enum import Enum
 from typing import Final
 
+from typing_extensions import deprecated
+
 LINUX_MEMORY_SLAB_STATE: Final = "linux.memory.slab.state"
 """
-The Linux Slab memory state.
+Deprecated: Replaced by `system.memory.linux.slab.state`.
 """
 
 
+@deprecated(
+    "The attribute linux.memory.slab.state is deprecated - Replaced by `system.memory.linux.slab.state`"
+)
 class LinuxMemorySlabStateValues(Enum):
     RECLAIMABLE = "reclaimable"
     """reclaimable."""